requirements.txt
flask
tailwind
numpy
scikit-image
pillow
sewar